app/requirements.txt
fastapi
uvicorn
python-multipart
pyyaml
Minio